+++ datacrawler/acceptance_tests/budget_finances_médico_social_test.py
@@ -12,15 +12,17 @@ from datacrawler.test_helpers import (
     NUMÉRO_FINESS_ENTITÉ_JURIDIQUE,
     NUMÉRO_FINESS_ÉTABLISSEMENT_MÉDICO_SOCIAL,
     base_de_données_test,
-    helios_ann_ca_ej_et_budget_et_finances_builder,
-    helios_ann_errd_ej_budget_et_finances_builder,
-    helios_ann_errd_ej_et_budget_et_finances_builder,
     mocked_logger,
     sauvegarde_les_indicateurs_budget_et_finances_en_base,
     sauvegarde_un_établissement_en_base,
     sauvegarde_une_date_de_mise_à_jour_de_fichier_source,
     sauvegarde_une_entité_juridique_en_base,
     supprime_les_données_des_tables,
+)
+from datacrawler.test_helpers.helios_builder import (
+    helios_ann_ca_ej_et_budget_et_finances_builder,
+    helios_ann_errd_ej_budget_et_finances_builder,
+    helios_ann_errd_ej_et_budget_et_finances_builder,
 )
